Preserve linked move status return width
The linked-Pokemon submenu narrows the move-status getter result into its 16-bit validity value before invoking the following move-validity helper. Declaring the getter's ABI-width return restores that explicit narrowing and the retail value lifetime.

Constraint: Keep the stored validity value 16-bit while matching the getter's register return contract.
Rejected: Widen the validity local | that removes the retail clrlwi normalization.
Confidence: high
Scope-risk: narrow
Directive: Do not narrow fn_8012640C's declaration without checking this caller's cross-call value lifetime.
Tested: ninja build/GC6E01/src/game/gbaCommunication_candidate_80096C48.o; ninja -j4 build/GC6E01/report.json; fn_80096FA0 fuzzy 85.08363 to 86.01091; git diff --check
Not-tested: Full game runtime

+++ src/game/gbaCommunication.c
@@ -3230,7 +3230,7 @@ void fn_80096FA0(u8* menu)
     extern void fn_80103484(s32, s32);
     extern void fn_80109C88(void*, u32);
     extern u8 fn_80123CD4(u32, s32);
-    extern u16 fn_8012640C(u32, u32, u32, s32);
+    extern u32 fn_8012640C(u32, u32, u32, s32);
 
     void* keyObject;
     u16 input;
